requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7

# HTTP Requests
requests==2.31.0

# File handling
Werkzeug==3.0.1

# Machine Learning (for Lambda functions)
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==11.0.0

# Additional utilities
python-dotenv==1.0.0
